requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-crispy-forms==1.9.2
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
